test(runner): extract shared scaffoldManagedRuntime test helper
